+++ subscription/api/blocking/src/main/java/org/occurrent/subscription/api/blocking/SubscriptionModel.java
@@ -91,4 +91,12 @@ public interface SubscriptionModel {
     default Subscription subscribe(String subscriptionId, Consumer<CloudEvent> action) {
         return subscribe(subscriptionId, null, StartAt.now(), action);
     }
+
+    /**
+     * Shutdown the subscription model and close all subscriptions (they can be resumed later if you start from a durable subscription position).
+     * A subscription model that is shutdown cannot be started again, since it closes resources such as database connections,
+     * thread pools etc.
+     */
+    default void shutdown() {
+    }
 }