+++ subscription/redis/spring/blocking-position-storage/src/test/java/org/occurrent/subscription/redis/spring/blocking/SpringRedisSubscriptionPositionStorageTest.java
@@ -36,6 +36,9 @@ import org.occurrent.functional.CheckedFunction;
 import org.occurrent.functional.Not;
 import org.occurrent.mongodb.timerepresentation.TimeRepresentation;
 import org.occurrent.retry.RetryStrategy;
+import org.occurrent.subscription.api.blocking.DelegatingSubscriptionModel;
+import org.occurrent.subscription.api.blocking.SubscriptionModel;
+import org.occurrent.subscription.api.blocking.SubscriptionModelLifeCycle;
 import org.occurrent.subscription.api.blocking.SubscriptionPositionStorage;
 import org.occurrent.subscription.blocking.durable.DurableSubscriptionModel;
 import org.occurrent.subscription.mongodb.spring.blocking.SpringMongoSubscriptionModel;
@@ -147,7 +150,7 @@ class SpringRedisSubscriptionPositionStorageTest {
 
         // When
         mongoEventStore.write("1", 0, serialize(nameDefined1));
-        redisSubscription.pauseSubscription(subscriberId);
+        pauseSubscription(redisSubscription, subscriberId);
         // The subscription is async so we need to wait for it
         await().atMost(ONE_SECOND).until(Not.not(state::isEmpty));
         mongoEventStore.write("2", 0, serialize(nameDefined2));
@@ -237,5 +240,14 @@ class SpringRedisSubscriptionPositionStorageTest {
         redisTemplate.setConnectionFactory(connectionFactory);
         redisTemplate.afterPropertiesSet();
         return redisTemplate;
+    }
+
+    private static void pauseSubscription(DelegatingSubscriptionModel subscriptionModel, String subscriberId) {
+        SubscriptionModel sm = subscriptionModel.getDelegatedSubscriptionModelRecursively();
+        if (sm instanceof SubscriptionModelLifeCycle) {
+            ((SubscriptionModelLifeCycle) sm).pauseSubscription(subscriberId);
+        } else {
+            throw new IllegalArgumentException("Cannot pause " + subscriberId);
+        }
     }
 }